// packages/text/types.ts
export interface TextPreset {
  mobileSize: number;
  desktopSize: number;
  mobileLineHeight: string;
  desktopLineHeight: string;
}

export enum Tag {
  H1 = 'h1',
  H2 = 'h2',
  H3 = 'h3',
  P = 'p',
  Span = 'span'
}

export enum FontWeight {
  Light = 'light',
  Regular = 'regular',
  Bold = 'bold'
}

export enum Preset {
  Text1 = 'text1',
  Text2 = 'text2',
  Text3 = 'text3',
  Text4 = 'text4',
  Text5 = 'text5',
  Text6 = 'text6',
  Text7 = 'text7',
  Text8 = 'text8',
  Text9 = 'text9',
  Text10 = 'text10'
}